+++ actions/upfront-tracks/UpdateTrack.ts
@@ -1,7 +1,6 @@
 'use server';
 
 import { db } from '@/db';
-import { getErrorMessage } from '@/lib/utils';
 
 type UpdateUpfrontMixParams = {
   upfrontId: string;
@@ -9,6 +8,7 @@ type UpdateUpfrontMixParams = {
   title?: string;
   artistId?: string;
   mixIds?: string[];
+  label?: string;
 };
 
 export async function updateUpfrontTrackWithMixes({
@@ -17,6 +17,7 @@ export async function updateUpfrontTrackWithMixes({
   title,
   artistId,
   mixIds,
+  label,
 }: UpdateUpfrontMixParams) {
   return await db.$transaction(async (tx) => {
     const track = await tx.upfrontTrack.findUnique({
@@ -37,27 +38,38 @@ export async function updateUpfrontTrackWithMixes({
     }
 
     if (artistId !== undefined) {
+      if (artistId === '') {
+        throw new Error('Artist field is required');
+      }
+
       await tx.upfrontTrack.update({
         where: { id: upfrontId },
         data: {
           artistId,
+          label,
         },
       });
     }
 
     if (mixIds !== undefined) {
+      if (mixIds.length === 0) {
+        throw new Error('Mixes field is required');
+      }
+
+      if (mixIds.some((mixId) => mixId === '')) {
+        throw new Error('Mixes field is required');
+      }
+
       await tx.upfrontMixTrack.deleteMany({
         where: { upfrontTrackId: upfrontId },
       });
 
-      if (mixIds.length > 0) {
-        await tx.upfrontMixTrack.createMany({
-          data: mixIds.map((mixId) => ({
-            upfrontTrackId: upfrontId,
-            mixId,
-          })),
-        });
-      }
+      await tx.upfrontMixTrack.createMany({
+        data: mixIds.map((mixId) => ({
+          upfrontTrackId: upfrontId,
+          mixId,
+        })),
+      });
     }
 
     return {
